chore: extract inline node utils to follow block file structure
