+++ packages/rich-text/src/Toolbar/components/EmbedEntityWidget.tsx
@@ -5,8 +5,8 @@ import { BLOCKS, INLINES } from '@contentful/rich-text-types';
 import { useContentfulEditor } from '../../ContentfulEditorProvider';
 import { isLinkActive } from '../../helpers/editor';
 import { isNodeTypeEnabled } from '../../helpers/validations';
-import { ToolbarEmbeddedEntityInlineButton } from '../../plugins/EmbeddedEntityInline';
 import { EmbeddedBlockToolbarIcon } from '../../plugins/shared/EmbeddedBlockToolbarIcon';
+import { EmbeddedInlineToolbarIcon } from '../../plugins/shared/EmbeddedInlineToolbarIcon';
 import { useSdkContext } from '../../SdkProvider';
 import { EmbeddedEntityDropdownButton } from './EmbeddedEntityDropdownButton';
 
@@ -50,7 +50,7 @@ export const EmbedEntityWidget = ({ isDisabled, canInsertBlocks }: EmbedEntityWi
         />
       )}
       {inlineEntryEmbedEnabled && (
-        <ToolbarEmbeddedEntityInlineButton
+        <EmbeddedInlineToolbarIcon
           isDisabled={!!isDisabled || isLinkActive(editor)}
           onClose={onCloseEntityDropdown}
         />
